[#22] 채팅 기능 구현 - 채팅 방의 채팅 내용 조회 기능 구현

+++ src/main/java/com/directors/infrastructure/jpa/chat/InmemoryChatRepository.java
@@ -15,7 +15,7 @@ public class InmemoryChatRepository implements ChatRepository {
     Map<Long, Chat> chatmap = new HashMap<>();
 
     @Override
-    public List<Chat> findChatListByRoomId(Long roomId, int offset, int size) {
+    public List<Chat> findChatListByRoomId(Long roomId, Long offset, Long size) {
         List<Chat> chatList = new ArrayList<>();
 
         List<Chat> chatByRoomId = chatmap.values()
